refactor(ownership): extract useOwnerProfilesForCar and rewrite use-ownerships onto it
Fetches ownerships for a car, derives owner ids, gates the dependent
owner-profiles query on whether any ownerships exist, and toasts errors for
both queries. use-ownerships.ts previously inlined this exact pattern.
